+++ apps/web/src/main.tsx
@@ -3,7 +3,10 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { App } from "./App";
 import { AppErrorBoundary } from "./components/AppErrorBoundary";
+import { requestBackendReadiness } from "./lib/backendReadiness";
 import "./styles.css";
+
+const initialBackendReadinessPromise = requestBackendReadiness();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
@@ -14,7 +17,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
           v7_relativeSplatPath: true,
         }}
       >
-        <App />
+        <App initialReadinessPromise={initialBackendReadinessPromise} />
       </BrowserRouter>
     </AppErrorBoundary>
   </React.StrictMode>,