+++ spec/index.ts
@@ -12,3 +12,4 @@ export * from "./events";
 export * from "./imageGeneration";
 export * from "./workflowLab";
 export * from "./spaceshipBoardState";
+export * from "./readiness";